requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-tenant-schemas==1.11.0
django-redis==5.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
redis==5.0.1
pika==1.3.2